% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agris.R
\name{agris}
\alias{agris}
\title{Agris Middlewares}
\usage{
agris()
}
\description{
Uses all agris middlewares.
}
\section{Middlewares}{

\itemize{
\item \code{\link[=use_content_security_policy]{use_content_security_policy()}}
\item \code{\link[=use_cross_origin_embedder_policy]{use_cross_origin_embedder_policy()}}
\item \code{\link[=use_cross_origin_opener_policy]{use_cross_origin_opener_policy()}}
\item \code{\link[=use_cross_origin_resource_policy]{use_cross_origin_resource_policy()}}
\item \code{\link[=use_dns_prefetch_control]{use_dns_prefetch_control()}}
\item \code{\link[=use_frame_options]{use_frame_options()}}
\item \code{\link[=use_hide_powered_by]{use_hide_powered_by()}}
\item \code{\link[=use_content_type_options]{use_content_type_options()}}
\item \code{\link[=use_xss_protection]{use_xss_protection()}}
\item \code{\link[=use_download_options]{use_download_options()}}
\item \code{\link[=use_strict_transport_security]{use_strict_transport_security()}}
}
}

