% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agris.R
\name{use_agris}
\alias{use_agris}
\title{All Agris Middlewares}
\usage{
use_agris()
}
\description{
All Agris Middlewares
}
\section{Middlewares}{

\itemize{
\item Nonce - \code{\link[=use_content_security_policy]{use_content_security_policy()}}
}
}

