% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csr.R
\name{use_content_security_policy}
\alias{use_content_security_policy}
\title{Use Content Security Policy}
\usage{
use_content_security_policy()
}
\description{
Adds relevant Content-Security-Policy headers.
}
\section{Directives}{

\itemize{
\item base-uri 'self';
\item block-all-mixed-content;
\item font-src 'self' https: data:;
\item form-action 'self';
\item frame-ancestors 'self';
\item img-src 'self' data:;
\item object-src 'none';
\item style-src 'self' https: 'unsafe-inline';
\item upgrade-insecure-requests
}
}

