% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cto.R
\name{use_content_type_options}
\alias{use_content_type_options}
\title{Content Type Options}
\usage{
use_content_type_options(value = "nosniff")
}
\arguments{
\item{value}{Value to set.}
}
\description{
Sets the \code{X-Content-Type-Options} to \code{nosniff} (default).
}
