% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coep.R
\name{use_cross_origin_embedder_policy}
\alias{use_cross_origin_embedder_policy}
\title{Cross Origin Embedder Policy}
\usage{
use_cross_origin_embedder_policy()
}
\description{
Sets the \code{Cross-Origin-Embedder-Policy} to \code{require-corp},
so the document can only load resources from the same origin,
or resources explicitly marked as loadable from another origin.
}
