% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coop.R
\name{use_cross_origin_opener_policy}
\alias{use_cross_origin_opener_policy}
\title{Cross Origin Opener Policy}
\usage{
use_cross_origin_opener_policy()
}
\description{
Sets the \code{Cross-Origin-Opener-Policy} to
\code{same-origin-allow-popups}.
}
