% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.R
\name{use_dns_prefetch_control}
\alias{use_dns_prefetch_control}
\title{DNS Prefetch Control}
\usage{
use_dns_prefetch_control(policy = c("on", "off"))
}
\arguments{
\item{policy}{Policy to set.}
}
\description{
Sets the \code{X-DNS-Prefetch-Control} header to \code{on} (default).
}
