% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{use_frame_options}
\alias{use_frame_options}
\title{Frame Options}
\usage{
use_frame_options(policy = c("DENY", "SAMEORIGIN"))
}
\arguments{
\item{policy}{Policy to set.}
}
\description{
Sets \code{X-Frame-Options} header to \code{DENY} (default).
}
