% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powered-by.R
\name{use_hide_powered_by}
\alias{use_hide_powered_by}
\title{Powered By}
\usage{
use_hide_powered_by(value = NULL)
}
\arguments{
\item{value}{Value to set the header to.}
}
\description{
Hides the \code{X-Powered-By} header.
Someone may want to exploit vulnerabilities of R or ambiorix,
hiding this provides less information to those people.
}
