% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oac.R
\name{use_origin_agent_cluster}
\alias{use_origin_agent_cluster}
\title{Origin Agent Cluster}
\usage{
use_origin_agent_cluster(value = "?1")
}
\arguments{
\item{value}{Value to set.}
}
\description{
Sets the \code{Origin-Agent-Cluster} to \code{true}.
Mechanism to allow web applications to isolate their origins.
}
