% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdp.R
\name{use_permitted_cross_domain_policies}
\alias{use_permitted_cross_domain_policies}
\title{Permitted Cross Domain Policies}
\usage{
use_permitted_cross_domain_policies(
  policy = c("none", "master-only", "by-content-type", "all")
)
}
\arguments{
\item{policy}{Policy to set.}
}
\description{
Sets the \code{X-Permitted-Cross-Domain-Policies} header to \code{none}.
Tells some clients (mostly Adobe products) your domain's policy
for loading cross-domain content.
}
