% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{use_referrer_policy}
\alias{use_referrer_policy}
\title{Referrer Policy}
\usage{
use_referrer_policy(
  policy = c("no-referrer", "no-referrer-when-downgrade", "origin",
    "origin-when-cross-origin", "same-origin", "strict-origin",
    "strict-origin-when-cross-origin", "unsafe-url")
)
}
\arguments{
\item{policy}{Policy to set.}
}
\description{
Sets the \code{Referrer-Policy} header to \code{no-referrer} (default).
Controls how much referrer information (sent with the Referer header)
should be included with requests.
}
