% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xss.R
\name{use_xss_protection}
\alias{use_xss_protection}
\title{XSS Protection}
\usage{
use_xss_protection(policy = "1; mode=block")
}
\arguments{
\item{policy}{Policy to set.}
}
\description{
Sets the \code{X-XSS-Protection} header to \verb{1; mode=block} to
enable XSS filtering and rather than sanitizing the page,
the browser will prevent rendering of the page if an attack is detected.
}
