% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csr.R
\name{use_content_security_policy}
\alias{use_content_security_policy}
\title{Use Content Security Policy}
\usage{
use_content_security_policy()
}
\description{
Adds relevant Content Security Policy headers.
}
\section{Changes}{

\itemize{
\item Adds \href{https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/nonce}{nonce}
to make each request unique.
}
}

