% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corp.R
\name{use_cross_origin_resource_policy}
\alias{use_cross_origin_resource_policy}
\title{Cross Origin Resource Policy}
\usage{
use_cross_origin_resource_policy()
}
\description{
Sets the \code{Cross-Origin-Resource-Policy} to \code{same-origin}.
}
